+++ lib/src/commands/gg_multi_do.dart
@@ -22,7 +22,6 @@ import 'do/push.dart';
 import 'do/publish.dart';
 import 'do/review.dart';
 import 'do/maintain.dart';
-import 'do/install_gitattributes.dart';
 
 /// Command to perform actions such as committing
 /// and pushing across ticket repositories.
@@ -52,7 +51,6 @@ class Do extends Command<void> {
     addSubcommand(DoConfigurePublishCommand(ggLog: ggLog));
     addSubcommand(DoReviewCommand(ggLog: ggLog));
     addSubcommand(MaintainCommand(ggLog: ggLog));
-    addSubcommand(DoInstallGitattributesCommand(ggLog: ggLog));
     addSubcommand(DoClaudeCommand(ggLog: ggLog));
     addSubcommand(AddDepsCommand(ggLog: ggLog));
     addSubcommand(AddCommand(ggLog: ggLog));