+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - refactor(gg_multi): replace do cancel-review with do review --abort
 
+### Removed
+
+- refactor(gg_multi): remove do install-git-attributes command
+
 ## 7.9.0 - 2026-08-02
 
 ### Added